requirements.txt
Django==5.0
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
liac-arff==2.5.0
gunicorn==21.2.0
